+++ CHANGELOG.md
@@ -9,6 +9,8 @@ https://github.com/jcorporation/myMPD/
 ### Changelog
 
 - Feat: Add option for Jukebox Autostart #1482
+- Feat: Manually trigger refill of the jukebox queue #1483
+- Feat: Configurable jukebox queue lengths #1484
 - Feat: Add option for default behavior on click on tag in browse view #1472
 - Feat: Optionally group songs with empty album tag in a special `Unknown Album` album #1472
 - Feat: Support large images